+++ MyQuizMobile/MyQuizMobile/MyQuizMobile/DataModel/Group.cs
@@ -1,6 +1,6 @@
 ﻿namespace MyQuizMobile.DataModel {
     public partial class Group {
-        public override int Id { get; set; }
+        public override long Id { get; set; }
         public string Title { get; set; }
         public string EnterGroupPin { get; set; }
     }
