+++ MyQuizMobile/MyQuizMobile/MyQuizMobile/DataModel/Item.cs
@@ -4,7 +4,7 @@ using PostSharp.Patterns.Model;
 namespace MyQuizMobile {
     [NotifyPropertyChanged]
     public abstract class Item {
-        public abstract int Id { get; set; }
+        public abstract long Id { get; set; }
         [JsonIgnore]
         public virtual string DisplayText { get; set; }
         [JsonIgnore]
